Make a small start with C99 features; misc related cleanup
We were already not C90-compliant as the code was using 'long long'
in a couple of places, so let's make it official now: we now
require a C99-ish compiler (gcc 2.95.3 is out, gcc 3.0+ should
still work).

The code now depends on the following C99 features:
- long long (as before)
- flexible array members (however, many instances of the 'struct
  hack' are still left to be converted to these; this change has
  some priority as the struct hack is considered UB)
- designated struct initialisers (init members by name)
- __func__ (name of current function)
- inline ('static inline' only for compat with gcc < 4.3)
- restrict (use of this optimisation keyword taken from #942 by
  @JohnoKing; not detailed further below, just see the diff)

I like testing ksh on really old systems as they have their own way
of exposing bugs. To avoid large diffs and to keep this historic
code base appropriately old-fashioned and (hopefully) compatible
with GCC 3.0+, I'm intending to keep it C90-ish with the use of
select C99 features such as the above. We also shouldn't assume an
OS with a C99-compliant libc yet.

This commit also silences some compiler warnings on macOS with
clang (CCFLAGS='-O0 -g -funsigned-char -Wall -Wsign-compare
-Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion').
The fixes I did to that end necessitated other fixes, and so they
cascaded and this commit got a bit large and messy; sorry about
that. (re: 00ecebd6, 95caccaa, 33e66eac)

src/cmd/INIT/hello.c:
- Rewrite the initial heuristic C compiler probe to require minimal
  C99 features. By @JohnoKing (taken from #942). I added a quick
  check for __func__.

src/cmd/INIT/mamake.c,
src/lib/libast/comp/assert.c,
src/lib/libast/std/assert.h,
src/lib/libast/misc/vmalloc.c:
- Remove code compiled when __STDC_VERSION__ < 199901L or for
  compilers that don't have __func__.

src/lib/libast/features/common:
- Remove obsolete __INLINE__ macro, defined as __inline on non-gcc
  _WIN32 only. Perhaps this was for UWIN. (re: 5c16b51f, 0ea97b9d)
- Remove feature tests for __func__ and the __FUNCTION__ fallback.

src/lib/libast/include/ast_std.h:
- Assume we have the inline keyword; remove empty macro fallback.

src/lib/libast/include/sfio.h,
src/lib/libast/sfio/sfflsbuf.c:
- Bump Sfio version to 20260907.
- When not building Sfio itself (!_BLD_sfio), declare static inline
  versions of the tiny sfput{d,l,u,m,c}, sfgetc, sf{d,l,u}len,
  sffileno, sfeof, sferror, sfclrerr, sfstacked, sfvalue, sfslen,
  and sfmaxr functions. These replace the macros or (theoretically,
  on non-gcc _WIN32) the __INLINE__ versions.
- Restore the int type for sfputc (it returns the character written
  or -1, so the ssize_t makes no sense). (re: fefe896e)
- For the same reason, restore int type to 2nd arg of _sfflsbuf().

src/lib/libast/sfio/_sf{clrerr,dlen,eof,error,fileno}.c,
src/lib/libast/sfio/_sfget{c,l,u}.c,
src/lib/libast/sfio/_sf{l,s,u}len.c,
src/lib/libast/sfio/_sf{putc,stacked,value}.c,
src/lib/libast/sfio/sfput{d,l,m,u}.c:
- Removed. All of these were either external-linkage fallbacks for
  non-static inline definitions (so were used on non-gcc _WIN32
  only) or were completely unused due to obsolescence.

src/lib/libast/sfio/sfflsbuf.c,
src/lib/libast/sfio/sfreserve.c:
- Restore the int type to the second argument to _sfflsbuf(): this is
  either negative (to return size of flushed data) or a byte,
  typecast to unsigned char, to add to the flushed data. Thus,
  ptrdiff_t makes no sense here.
- In sfreserve(), fix a bug exposed by a compiler warning triggered
  by the change above. In one of the SFFLSBUF() macro expansions,
  which calls _sfflsbuf(), the second argument was a data size,
  iosz. This is incorrect (see above) and would have caused a
  garbage byte to be added to the flushed data. Change to -1, which
  is probably what was intended given other such calls that use -1.

src/lib/libast/sfio/sfvprintf.c:
- Delete small fallback macros that were only compiled on Amdahl
  UTS, a Unix system for mainframes defunct since 2002 or before.

src/lib/libcmd/cat.c:
- regress(): The static inline function version of sfvalue() is not
  an lvalue, so can't be assigned to; since this is a hack anyway,
  just use the internal __sf_value() macro definition as an lvalue.

src/lib/libast/include/sfio_t.h,
src/lib/libast/misc/state.c,
src/lib/libast/comp/localeconv.c:
- Use designated struct initialisers.

src/lib/libast/regex/regcomp.c:
- Work around a compiler warning on macOS where MB_CUR_MAX is of
  type int (contrarily to POSIX which specifies size_t).
  Typecasting the right-hand side to the left-hand type (ptrdiff_t)
  works regardless of the type of MB_CUR_MAX.

src/lib/libcmd/stty.c:
- Silence compiler warning on macOS and FreeBSD, where the
  sethostname() definition differs: even in 2026, it still uses an
  int as the second (name length) argument instead of size_t. All
  other systems seem to have switched to size_t by now.
- Fix a bug: add braces around error()/UNREACHABLE() as they will
  usually be conditional upon the preceding 'if', though the
  somewhat ridiculous preprocessor stuff obscures this.

src/lib/libcmd/join.c,
src/cmd/ksh93/sh/macro.c:
- Introduce new variables instead of reusing existing ones, so each
  one can have an appropriate type. This fixes compiler warnings.

src/lib/libast/features/api:
- Bump API version to 20260907.

+++ src/cmd/ksh93/sh/tdump.c
@@ -63,7 +63,7 @@ static ssize_t p_tree(const Shnode_t *t)
 			return p_comarg((struct comnod*)t);
 		case TSETIO:
 		case TFORK:
-			if(sfputu(outfile,t->fork.forkline)<0)
+			if(sfputu(outfile,(uintmax_t)t->fork.forkline)<0)
 				return -1;
 			if(p_tree(t->fork.forktre)<0)
 				return -1;
@@ -96,11 +96,11 @@ static ssize_t p_tree(const Shnode_t *t)
 				return -1;
 			return p_tree(t->lst.lstrit);
 		case TARITH:
-			if(sfputu(outfile,t->ar.arline)<0)
+			if(sfputu(outfile,(uintmax_t)t->ar.arline)<0)
 				return -1;
 			return p_arg(t->ar.arexpr);
 		case TFOR:
-			if(sfputu(outfile,t->for_.forline)<0)
+			if(sfputu(outfile,(uintmax_t)t->for_.forline)<0)
 				return -1;
 			if(p_tree(t->for_.fortre)<0)
 				return -1;
@@ -108,13 +108,13 @@ static ssize_t p_tree(const Shnode_t *t)
 				return -1;
 			return p_tree((Shnode_t*)t->for_.forlst);
 		case TSW:
-			if(sfputu(outfile,t->sw.swline)<0)
+			if(sfputu(outfile,(uintmax_t)t->sw.swline)<0)
 				return -1;
 			if(p_arg(t->sw.swarg)<0)
 				return -1;
 			return p_switch(t->sw.swlst);
 		case TFUN:
-			if(sfputu(outfile,t->funct.functline)<0)
+			if(sfputu(outfile,(uintmax_t)t->funct.functline)<0)
 				return -1;
 			if(p_string(t->funct.functnam)<0)
 				return -1;
@@ -122,7 +122,7 @@ static ssize_t p_tree(const Shnode_t *t)
 				return -1;
 			return p_tree((Shnode_t*)t->funct.functargs);
 		case TTST:
-			if(sfputu(outfile,t->tst.tstline)<0)
+			if(sfputu(outfile,(uintmax_t)t->tst.tstline)<0)
 				return -1;
 			if((t->tre.tretyp&TPAREN)==TPAREN)
 				return p_tree(t->lst.lstlef);
@@ -159,7 +159,7 @@ static ssize_t p_arg(const struct argnod *arg)
 		sfputc(outfile,arg->argflag);
 		if(fp)
 		{
-			sfputu(outfile,fp->fortyp);
+			sfputu(outfile,(uintmax_t)fp->fortyp);
 			p_tree(fp->fortre);
 		}
 		else if(n==0 && (arg->argflag&ARG_EXP) && arg->argchn.ap)
@@ -207,7 +207,7 @@ static ssize_t p_comarg(const struct comnod *com)
 		p_arg(com->comarg.ap);
 	else
 		p_comlist(com->comarg.dp);
-	return sfputu(outfile,com->comline);
+	return sfputu(outfile,(uintmax_t)com->comline);
 }
 
 static ssize_t p_comlist(const struct dolnod *dol)
